refactor: move to `common/websockets`
